Add a --no-summary option for cmdline output
Currently, the cmdline output is pretty regular, with the exception of the
summary information ("No suggestions", "n suggestions."). In order to simplify
a bit the parsing of cmdline output (where full power of json is not needed), a
--no-summary flag is added that skips this last bit. This fixes #67.

Tests are added as well, although this test suite format is a bit strange and
I'm not sure I did the right thing.