Stop unsafely binding a new variable
Currently, hlint will suggest turning `zipWith (+) xs (repeat x)` into `map (\ x -> (+) x x) xs`, which is wrong. The problem is that it's not safe to bind a new variable and use user-provided expressions in the same scope. This change results in it suggesting the correct `map (+ x) xs` instead.