Warn on partial functions.
The warnings are for partial functions in the `base` package.  There are two groups that are disabled by default: `partial` for functions that should be relatively easy to avoid, and `partial-strict` for functions whose use may not be feasible to avoid.

Suggestions have been added for partial functions in the `partial` group where they can be reasonably added.  For those that must necessarily remain partial, `error _` is used as the value for the undefined cases.  This both serves as a signal that the case remains undefined, and forces it to be rewritten appropriately by being a syntax error.